requirements.txt
anyio==3.6.1
click==8.1.3
debugpy==1.6.3
fastapi==0.85.0
h11==0.13.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.12.0
pydantic==1.10.2
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
zipp==3.8.1
